refactor: 对齐上游评审约束 —— 类型集中 types.ts、规则去重
对照上游 PR #38 三轮评审的约束口径自查本分支,修正三处同类问题:

- 「前端类型集中于 types.ts」(上游对 ProjectKind/AiVendor 定义在 utils 判 P2 硬性):
  跨文件共享的 PreviewRun/PreviewGrid/PreviewPaletteOptions 挪入 types.ts;
  PreviewCellLike 等结构性 buffer 接口按 terminalSnapshot.ts 未被点名的先例留在
  panePreview.ts(它们是对 xterm API 形状的测试性抽象,不是领域类型)
- 「同一规则散落多处」(上游对 App.tsx/store.ts 重复实现托盘优先级判 P3):
  预览浮层私有的 STATUS_RANK 删除,改用 store.ts 已导出的 STATUS_PRIORITY
  (ActivityBar 同款用法);pane 显示名口径(customTitle > 远程连接名 > shellName)
  从 PaneGroup 局部函数与浮层 labelOf 两处实现收敛为 remoteProject.ts 的
  paneDisplayLabel,两处调用方共用
- git diff --check 全程干净(上游曾对 EOF 空行告警)

store.ts 唯一全局状态约束核对无违反:浮层仅组件级 state + useAppStore 订阅,
未新建模块级可订阅状态
