+++ src/types.ts
@@ -321,6 +321,29 @@ export interface PaneState {
   attention?: boolean;
 }
 
+// === pane 预览缩略图(panePreview.ts 提取 → panePreviewCanvas.ts 绘制) ===
+
+/** 同色连续字符段;col 为起始列,绘制定位 x = col × cellW */
+export interface PreviewRun {
+  col: number;
+  text: string;
+  color: string;
+}
+
+export interface PreviewGrid {
+  cols: number;
+  rows: number;
+  /** 每视口行的 runs;空白不产生 run */
+  lines: PreviewRun[][];
+}
+
+export interface PreviewPaletteOptions {
+  /** ANSI 16 色(black..white, brightBlack..brightWhite),来自终端主题 */
+  palette16: string[];
+  /** 默认前景色(theme.foreground) */
+  foreground: string;
+}
+
 // === AI 会话 ===
 
 export interface AiSession {
